hyperliquid-builders-dashboard/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
grpcio==1.68.1
grpcio-tools==1.68.1
python-dotenv==1.0.1
websockets==14.1
